backend/requirements.txt
Flask
Flask-SocketIO
gevent-websocket
SQLAlchemy
PyMySQL
Flask-Cors
gunicorn
python-dotenv
google-generativeai
Flask-Login